# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Panacea Socios API",
    "description": "API para gesti\u00f3n de socios, sus preferencias y registro de ventas de Panacea. Permite registrar clientes, capturar sus respuestas a preguntas clave y llevar un historial de compras para futuras promociones y regalos.",
    "contact": {
      "name": "Panacea",
      "email": "info@panacea.com"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/socios/": {
      "get": {
        "tags": [
          "Socios"
        ],
        "summary": "Listar socios",
        "description": "Retorna la lista paginada de socios. Requiere admin.",
        "operationId": "list_socios_socios__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "dni",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dni"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SocioResponse"
                  },
                  "title": "Response List Socios Socios  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Socios"
        ],
        "summary": "Crear socio",
        "description": "Crea un nuevo socio. Requiere admin. DNI y email deben ser \u00fanicos.",
        "operationId": "create_socio_socios__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocioCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/socios/me": {
      "get": {
        "tags": [
          "Socios"
        ],
        "summary": "Obtener perfil propio",
        "description": "Retorna el perfil del socio autenticado.",
        "operationId": "get_me_socios_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocioResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/socios/{socio_id}": {
      "get": {
        "tags": [
          "Socios"
        ],
        "summary": "Obtener socio",
        "description": "Retorna los datos de un socio por su ID. Requiere autenticaci\u00f3n.",
        "operationId": "get_socio_socios__socio_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "socio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Socio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Socios"
        ],
        "summary": "Actualizar socio",
        "description": "Actualiza parcialmente los datos de un socio. Requiere admin.",
        "operationId": "update_socio_socios__socio_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "socio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Socio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocioUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Socios"
        ],
        "summary": "Eliminar socio",
        "description": "Elimina un socio y sus preguntas asociadas. Requiere admin.",
        "operationId": "delete_socio_socios__socio_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "socio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Socio Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SocioCreate": {
        "properties": {
          "nombre_apellido": {
            "type": "string",
            "maxLength": 200,
            "title": "Nombre Apellido",
            "examples": [
              "Juan P\u00e9rez"
            ]
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono",
            "examples": [
              "+54 9 11 1234-5678"
            ]
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "examples": [
              "juan@example.com"
            ]
          },
          "fecha_nacimiento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Nacimiento",
            "examples": [
              "1990-05-15"
            ]
          },
          "dni": {
            "type": "string",
            "maxLength": 20,
            "title": "Dni",
            "examples": [
              "30123456"
            ]
          }
        },
        "type": "object",
        "required": [
          "nombre_apellido",
          "dni"
        ],
        "title": "SocioCreate"
      },
      "SocioResponse": {
        "properties": {
          "nombre_apellido": {
            "type": "string",
            "maxLength": 200,
            "title": "Nombre Apellido",
            "examples": [
              "Juan P\u00e9rez"
            ]
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono",
            "examples": [
              "+54 9 11 1234-5678"
            ]
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "examples": [
              "juan@example.com"
            ]
          },
          "fecha_nacimiento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Nacimiento",
            "examples": [
              "1990-05-15"
            ]
          },
          "dni": {
            "type": "string",
            "maxLength": 20,
            "title": "Dni",
            "examples": [
              "30123456"
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "nombre_apellido",
          "dni",
          "id"
        ],
        "title": "SocioResponse"
      },
      "SocioUpdate": {
        "properties": {
          "nombre_apellido": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre Apellido"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "fecha_nacimiento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Nacimiento"
          },
          "dni": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Dni"
          }
        },
        "type": "object",
        "title": "SocioUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  }
}